Tests: Added test for definition of local variables in adjacent scopes
"super" is defined as a local variable with scope_depth_ == 1. In
earlier implementations, "super" was added as a local variable but not
removed once a class statement had finished compilation. If a subsequent
local scope was used outside of the class statement, the variable
pointed to by super, i.e. the superclass, would end up being referenced.
This test verifies that this behaviour does not occur.